docs/builder/requirements.frozen.txt
markdown==3.4.4
beautifulsoup4==4.12.2
Jinja2==3.1.2
PyYAML==6.0.1
soupsieve==2.5
MarkupSafe==2.1.3
lxml==4.9.3
html5lib==1.1